Clean code
git-svn-id: svn://pc-lab14/SVN/cloudify/trunk/cloudify@114817 eb64e737-3616-4df0-8941-5ee2ae88103d

+++ restful/src/main/java/org/cloudifysource/rest/controllers/ServiceController.java
@@ -1338,17 +1338,17 @@ public class ServiceController {
 
 		CapacityRequirementsConfig minCapacity = 
 			new CapacityRequirementsConfigurer()
-			.memoryCapacity((int)(service.getMinAllowedInstances() * externalProcessMemoryInMB), MemoryUnit.MEGABYTES)
+			.memoryCapacity((service.getMinAllowedInstances() * externalProcessMemoryInMB), MemoryUnit.MEGABYTES)
 			.create();
 
 		CapacityRequirementsConfig initialCapacity = 
 			new CapacityRequirementsConfigurer()
-			.memoryCapacity((int) (service.getNumInstances() * externalProcessMemoryInMB), MemoryUnit.MEGABYTES)
+			.memoryCapacity((service.getNumInstances() * externalProcessMemoryInMB), MemoryUnit.MEGABYTES)
 			.create();
 
 		CapacityRequirementsConfig maxCapacity = 
 			new CapacityRequirementsConfigurer()
-			.memoryCapacity((int)(service.getMaxAllowedInstances() * externalProcessMemoryInMB), MemoryUnit.MEGABYTES)
+			.memoryCapacity((service.getMaxAllowedInstances() * externalProcessMemoryInMB), MemoryUnit.MEGABYTES)
 			.create();
 
 		AutomaticCapacityScaleConfig scaleConfig =
